# kratos_example/helloworld/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Todo API
    version: 0.0.1
paths:
    /v1/todo:
        get:
            tags:
                - Todo
            operationId: Todo_ListTodo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.helloworld.v1.ListTodoReply'
        post:
            tags:
                - Todo
            operationId: Todo_CreateTodo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.helloworld.v1.CreateTodoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.helloworld.v1.CreateTodoReply'
    /v1/todo/{id}:
        get:
            tags:
                - Todo
            operationId: Todo_GetTodo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.helloworld.v1.GetTodoReply'
        put:
            tags:
                - Todo
            operationId: Todo_UpdateTodo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.helloworld.v1.UpdateTodoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.helloworld.v1.UpdateTodoReply'
        delete:
            tags:
                - Todo
            operationId: Todo_DeleteTodo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.helloworld.v1.DeleteTodoReply'
components:
    schemas:
        api.helloworld.v1.CreateTodoReply:
            type: object
            properties:
                todo:
                    $ref: '#/components/schemas/api.helloworld.v1.TodoMessage'
        api.helloworld.v1.CreateTodoRequest:
            type: object
            properties:
                massageInfo:
                    type: string
        api.helloworld.v1.DeleteTodoReply:
            type: object
            properties: {}
        api.helloworld.v1.GetTodoReply:
            type: object
            properties:
                todo:
                    $ref: '#/components/schemas/api.helloworld.v1.TodoMessage'
        api.helloworld.v1.ListTodoReply:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.helloworld.v1.TodoMessage'
        api.helloworld.v1.TodoMessage:
            type: object
            properties:
                id:
                    type: string
                massageInfo:
                    type: string
                states:
                    type: boolean
        api.helloworld.v1.UpdateTodoReply:
            type: object
            properties: {}
        api.helloworld.v1.UpdateTodoRequest:
            type: object
            properties:
                id:
                    type: string
                massageInfo:
                    type: string
                states:
                    type: boolean
tags:
    - name: Todo